zram: Rate limit memory allocation errors
If an error occurs allocating memory for zram we will not be able to
fullfill the request to store the page in zram. The swap subsystem still
continues to try to swap out pages to zram even when this error occurs
since there is currently no facility to stop the swap subsystem from
swapping out during such errors. This can cause the system to be
overflowed with logging errors.

Reduce the amount of logging to prevent the kernel log from being filled
with these error messages.

Change-Id: I70b65b8db548a7b57a6ee7fc70cd23dd0f5f5d89
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
Signed-off-by: karthikkm007 <jacksparrow941994@gmail.com>